docs: more details about theme
